requirements.txt
PyMuPDF
Pillow
indic-transliteration
streamlit
python-dotenv
pytesseract 